fet: management list variant
