+++ src/i18n/en/Employee.PaymentMethod.json
@@ -49,6 +49,8 @@
   "priority_two": "{{count}}nd",
   "priority_few": "{{count}}rd",
   "priority_other": "{{count}}th",
+  "managementTitle": "Payment",
+  "paymentMethodLabel": "Payment method",
   "validations": {
     "percentageError": "If payment method amount is split by Percentage, all split amounts must add up to exactly 100.",
     "percentageErrorWithTotal": "Splits must total 100%. Currently {{total}}%.",
